requirements.txt
pygame-ce==2.3.1
PyOpenGL==3.1.7
